view/view_test.go
package view

import (
	"github.com/stretchr/testify/assert"
	"html/template"
	"testing"
)

func TestView_Render(t *testing.T) {
	var v *View
	var content template.HTML
	v = New()
	content = v.Render("html/tpl.html", map[string]interface{}{
		"Title":   "ThinkGo",
		"Message": "Hello ThinkGo !",
	})

	assert.Contains(t, content, "<h2>Hello ThinkGo !</h2>")
	assert.Contains(t, content, "<title>ThinkGo</title>")

	v = New()
	v.ParseGlob("html/*")
	content = v.Render("tpl.html", map[string]interface{}{
		"Title":   "ThinkGo",
		"Message": "Hello ThinkGo !",
	})

	assert.Contains(t, content, "<h2>Hello ThinkGo !</h2>")
	assert.Contains(t, content, "<title>ThinkGo</title>")
}
